+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Always pass the `productId` upon item click while submitting the click metric.
+
 ## [2.1.0] - 2025-01-07
 
 ### Changed