+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.1.1] - 2025-01-13
+
 ### Fixed
 
 - Always pass the `productId` upon item click while submitting the click metric.